Editor-UI fixes
